requirements.txt
aiohttp==3.9.5
google-generativeai==0.5.4
lxml_html_clean==0.1.1
newspaper3k==0.2.8
pillow==10.3.0
discord.py==2.3.2
PyNaCl==1.5.0
supabase==2.5.0
cohere==5.5.6
python-dotenv==1.0.1
